Fix translates module
