+++ src/app/shared/shared.module.ts
@@ -1,12 +1,12 @@
 import { ModuleWithProviders, NgModule } from '@angular/core';
 
 import { LayoutsModule } from './layouts/layouts.module';
-import { TranslateModule } from './translate/translate.module';
+import { TranslatesModule } from './translates/translates.module';
 
 @NgModule({
   exports: [
     LayoutsModule,
-    TranslateModule
+    TranslatesModule
   ],
   providers: []
 })
